+++ website/notifications.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, request, render_template, redirect, url_for
 from flask_login import login_required, current_user
-from models import db, Notification
+from .models import db, Notification
 from datetime import datetime
 
 notifications_blueprint = Blueprint("notifications", __name__)
