+++ website/seed_rooms.py
@@ -1,8 +1,9 @@
-from app import app, db
-from models import Room
+from website import db
+from . import website
+from .models import Room
 
 def seed_rooms():
-    with app.app_context():
+    with website.app_context():
         # Check if rooms already exist
         if Room.query.first():
             print("Rooms already exist in database")
