+++ website/wsgi.py
@@ -1,4 +1,4 @@
-from app import app
+from website import app
 
 # Expose the Flask app as a WSGI application
 application = app
